+++ cas-server-integration-memcached/pom.xml
@@ -89,6 +89,12 @@
             <artifactId>hibernate-validator</artifactId>
             <scope>test</scope>
         </dependency>
+
+        <dependency>
+            <groupId>com.ryantenney.metrics</groupId>
+            <artifactId>metrics-spring</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <properties>
