+++ cas-server-integration-memcached/src/main/java/org/jasig/cas/monitor/MemcachedMonitor.java
@@ -45,7 +45,6 @@ public class MemcachedMonitor extends AbstractCacheMonitor {
     @Qualifier("memcachedClient")
     private MemcachedClientIF memcachedClient;
 
-
     /**
      * Instantiates a new Memcached monitor.
      */
